+++ GameServer/EnemyNomalSkeleton.cpp
@@ -125,5 +125,5 @@ void EnemyNomalSkeleton::OnReward()
 		return;
 	}
 
-	instigated->GetLevelComponent().AddExperience(Random::GetNormalDistribution(40, 0.8));
+	instigated->GetLevelComponent().AddExperience(static_cast<int32>(Random::GetNormalDistribution(40, 0.8)));
 }
